+++ sample-resource-plugin/src/main/java/org/opensearch/sample/resource/actions/rest/create/UpdateResourceRequest.java
@@ -18,6 +18,7 @@ import org.opensearch.core.common.io.stream.StreamOutput;
 import org.opensearch.sample.SampleResource;
 
 import static org.opensearch.sample.utils.Constants.RESOURCE_INDEX_NAME;
+import static org.opensearch.sample.utils.Constants.RESOURCE_TYPE;
 
 /**
  * Request object for UpdateResource transport action
@@ -57,6 +58,11 @@ public class UpdateResourceRequest extends ActionRequest implements DocRequest {
 
     public String getResourceId() {
         return this.resourceId;
+    }
+
+    @Override
+    public String type() {
+        return RESOURCE_TYPE;
     }
 
     @Override